requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch
pandas
openpyxl
jinja2
python-multipart
aiofiles
